disp: new option --fps to maintain playback rate
Add a new rate control mechanism to the generic display, to allow forcing the
playback rate from command line.

We add rate control calls in the generic display functions disp_post_buffer()
and disp_post_vid_buffer(). As this is in the generic display part, x11 and kms
(and future) displays benefit from that mechanism.

We use the mark() and usleep() functions for the rate control implementation,
and we do also filter a bit the sleep duration, to avoid being too "choppy".

Signed-off-by: Vincent Stehlé <v-stehle@ti.com>